lekce_9/pyproject.toml
[project]
name = "lekce_9"
version = "0.1.0"
description = "Universal LangChain Agent with MCP Server Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.9.0",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
